image to image support for UI Image and RawImage
works with readable, uncompressed textures
