+++ Assets/Stable-Diffusion-Unity-Integration/Input Image.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 31317b8dcb4d62045ad8e7c81001fad8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
