+++ Assets/Stable-Diffusion-Unity-Integration/Scripts/Editor/StableDiffusionImage2ImageEditor.cs
@@ -1,0 +1,31 @@
+using UnityEditor;
+using UnityEngine;
+
+/// <summary>
+/// Custom Inspector to add a models list as a drop-down selection UI 
+/// and add a "Generate" button for the StableDiffusionImage.
+/// </summary>
+[CustomEditor(typeof(StableDiffusionImage2Image))]
+public class StableDiffusionImage2ImageEditor : Editor
+{
+    public override void OnInspectorGUI()
+    {
+        EditorGUILayout.HelpBox("Input textures should be uncompressed and readable", MessageType.Info);
+
+        base.OnInspectorGUI();
+
+        StableDiffusionImage2Image myComponent = (StableDiffusionImage2Image)target;
+
+        // Draw the drop-down list for the Samplers list
+        myComponent.selectedSampler = EditorGUILayout.Popup("Sampler", myComponent.selectedSampler, myComponent.samplersList);
+
+        // Draw the drop-down list for the Models list
+        myComponent.selectedModel = EditorGUILayout.Popup("Model", myComponent.selectedModel, myComponent.modelsList);
+
+        // Apply the changes to the serialized object
+        serializedObject.ApplyModifiedProperties();
+
+        if (GUILayout.Button("Generate"))
+            myComponent.Generate();
+    }
+}
