+++ Assets/Stable-Diffusion-Unity-Integration/Scripts/Editor/StableDiffusionImage2ImageEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ef2486aabc75eb48a62c0d4bb8051bf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
