--- Assets/Stable-Diffusion-Unity-Integration/Scripts/Editor/StableDiffusionMaterialEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 285b9067714414e499dbdff9a234b312
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
