+++ Assets/Stable-Diffusion-Unity-Integration/Scripts/Editor/StableDiffusionText2ImageEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 98177daecf2a6ae4e973eb5975e11ecc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
