+++ Assets/Stable-Diffusion-Unity-Integration/Scripts/Editor/StableDiffusionText2MaterialEditor.cs
@@ -1,0 +1,29 @@
+using UnityEditor;
+using UnityEngine;
+
+/// <summary>
+/// Custom Inspector to add a models list as a drop-down selection UI 
+/// and add a "Generate" button for the StableDiffusionMaterial.
+/// </summary>
+[CustomEditor(typeof(StableDiffusionText2Material))]
+public class StableDiffusionText2MaterialEditor : Editor
+{
+    public override void OnInspectorGUI()
+    {
+        base.OnInspectorGUI();
+
+        StableDiffusionText2Material myComponent = (StableDiffusionText2Material)target;
+
+        // Draw the drop-down list for the Samplers list
+        myComponent.selectedSampler = EditorGUILayout.Popup("Sampler", myComponent.selectedSampler, myComponent.samplersList);
+
+        // Draw the drop-down list for the Models list
+        myComponent.selectedModel = EditorGUILayout.Popup("Model", myComponent.selectedModel, myComponent.modelsList);
+        
+        // Apply the changes to the serialized object
+        serializedObject.ApplyModifiedProperties();
+
+        if (GUILayout.Button("Generate"))
+            myComponent.Generate();
+    }
+}
