--- Assets/Stable-Diffusion-Unity-Integration/Scripts/StableDiffusionImage.cs
@@ -1,350 +1,0 @@
-using Newtonsoft.Json;
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.IO;
-using System.Net;
-using UnityEngine;
-using System.Linq;
-using UnityEngine.UI;
-using System.Threading.Tasks;
-
-#if UNITY_EDITOR
-using UnityEditor;
-using UnityEditor.SceneManagement;
-#endif
-
-/// <summary>
-/// Component to help generate a UI Image or RawImage using Stable Diffusion.
-/// </summary>
-[ExecuteAlways]
-public class StableDiffusionImage : StableDiffusionGenerator
-{
-    [ReadOnly]
-    public string guid = "";
-    
-    public string prompt;
-    public string negativePrompt;
-
-    /// <summary>
-    /// List of samplers to display as Drop-Down in the inspector
-    /// </summary>
-    [SerializeField]
-    public string[] samplersList
-    {
-        get
-        {
-            if (sdc == null)
-                sdc = GameObject.FindObjectOfType<StableDiffusionConfiguration>();
-            return sdc.samplers;
-        }
-    }
-    /// <summary>
-    /// Actual sampler selected in the drop-down list
-    /// </summary>
-    [HideInInspector]
-    public int selectedSampler = 0;
-
-    public int width = 512;
-    public int height = 512;
-    public int steps = 90;
-    public float cfgScale = 7;
-    public long seed = -1;
-
-    public long generatedSeed = -1;
-
-    string filename = "";
-
-
-
-    /// <summary>
-    /// List of models to display as Drop-Down in the inspector
-    /// </summary>
-    [SerializeField]
-    public string[] modelsList
-    {
-        get
-        {
-            if (sdc == null)
-                sdc = GameObject.FindObjectOfType<StableDiffusionConfiguration>();
-            return sdc.modelNames;
-        }
-    }
-    /// <summary>
-    /// Actual model selected in the drop-down list
-    /// </summary>
-    [HideInInspector]
-    public int selectedModel = 0;
-
-
-    /// <summary>
-    /// On Awake, fill the properties with default values from the selected settings.
-    /// </summary>
-    void Awake()
-    {
-#if UNITY_EDITOR
-        if (width < 0 || height < 0)
-        {
-            StableDiffusionConfiguration sdc = GameObject.FindObjectOfType<StableDiffusionConfiguration>();
-            if (sdc != null)
-            {
-                SDSettings settings = sdc.settings;
-                if (settings == null)
-                {
-
-                    width = settings.width;
-                    height = settings.height;
-                    steps = settings.steps;
-                    cfgScale = settings.cfgScale;
-                    seed = settings.seed;
-                    return;
-                }
-            }
-
-            width = 512;
-            height = 512;
-            steps = 50;
-            cfgScale = 7;
-            seed = -1;
-        }
-#endif
-    }
-
-
-    void Update()
-    {
-#if UNITY_EDITOR
-        // Clamp image dimensions values between 128 and 2048 pixels
-        if (width < 128) width = 128;
-        if (height < 128) height = 128;
-        if (width > 2048) width = 2048;
-        if (height > 2048) height = 2048;
-
-        // If not setup already, generate a GUID (Global Unique Identifier)
-        if (guid == "")
-            guid = Guid.NewGuid().ToString();
-#endif
-    }
-
-    // Internally keep tracking if we are currently generating (prevent re-entry)
-    bool generating = false;
-
-    /// <summary>
-    /// Callback function for the inspector Generate button.
-    /// </summary>
-    public void Generate()
-    {
-        // Start generation asynchronously
-        if (!generating && !string.IsNullOrEmpty(prompt))
-        {
-            StartCoroutine(GenerateAsync());
-        }
-    }
-
-    /// <summary>
-    /// Setup the output path and filename for image generation
-    /// </summary>
-    void SetupFolders()
-    {
-        // Get the configuration settings
-        if (sdc == null)
-            sdc = GameObject.FindObjectOfType<StableDiffusionConfiguration>();
-
-        try
-        {
-            // Determine output path
-            string root = Application.dataPath + sdc.settings.OutputFolder;
-            if (root == "" || !Directory.Exists(root))
-                root = Application.streamingAssetsPath;
-            string mat = Path.Combine(root, "SDImages");
-            filename = Path.Combine(mat, guid + ".png");
-
-            // If folders not already exists, create them
-            if (!Directory.Exists(root))
-                Directory.CreateDirectory(root);
-            if (!Directory.Exists(mat))
-                Directory.CreateDirectory(mat);
-
-            // If the file already exists, delete it
-            if (File.Exists(filename))
-                File.Delete(filename);
-        }
-        catch (Exception e)
-        {
-            Debug.LogError(e.Message + "\n\n" + e.StackTrace);
-        }
-    }
-
-    IEnumerator GenerateAsync()
-    {
-        generating = true;
-
-        SetupFolders();
-        
-        // Set the model parameters
-        yield return sdc.SetModelAsync(modelsList[selectedModel]);
-
-        // Generate the image
-        HttpWebRequest httpWebRequest = null;
-        try
-        {
-            // Make a HTTP POST request to the Stable Diffusion server
-            httpWebRequest = (HttpWebRequest)WebRequest.Create(sdc.settings.StableDiffusionServerURL + sdc.settings.TextToImageAPI);
-            httpWebRequest.ContentType = "application/json";
-            httpWebRequest.Method = "POST";
-
-            // Send the generation parameters along with the POST request
-            using (var streamWriter = new StreamWriter(httpWebRequest.GetRequestStream()))
-            {
-                SDParamsIn sd = new SDParamsIn();
-                sd.prompt = prompt;
-                sd.negative_prompt = negativePrompt;
-                sd.steps = steps;
-                sd.cfg_scale = cfgScale;
-                sd.width = width;
-                sd.height = height;
-                sd.seed = seed;
-                sd.tiling = false;
-
-                if (selectedSampler >= 0 && selectedSampler < samplersList.Length)
-                    sd.sampler_name = samplersList[selectedSampler];
-
-                // Serialize the input parameters
-                string json = JsonConvert.SerializeObject(sd);
-
-                // Send to the server
-                streamWriter.Write(json);
-            }
-        }
-        catch (Exception e)
-        {
-            Debug.LogError(e.Message + "\n\n" + e.StackTrace);
-        }
-
-        // Read the output of generation
-        if (httpWebRequest != null)
-        {
-            // Wait that the generation is complete before procedding
-            Task<WebResponse> t = httpWebRequest.GetResponseAsync();
-            while (!t.IsCompleted)
-            {
-                UpdateGenerationProgress();
-                yield return new WaitForSeconds(0.5f);
-            }
-            var httpResponse = t.Result;
-
-            using (var streamReader = new StreamReader(httpResponse.GetResponseStream()))
-            {
-                // Decode the response as a JSON string
-                string result = streamReader.ReadToEnd();
-
-                // Deserialize the JSON string into a data structure
-                SDResponse json = JsonConvert.DeserializeObject<SDResponse>(result);
-
-                // If no image, there was probably an error so abort
-                if (json.images == null || json.images.Length == 0)
-                {
-                    Debug.LogError("No image was return by the server. This should not happen. Verify that the server is correctly setup.");
-
-                    generating = false;
-#if UNITY_EDITOR
-                    EditorUtility.ClearProgressBar();
-#endif
-                    yield break;
-                }
-
-                // Decode the image from Base64 string into an array of bytes
-                byte[] imageData = Convert.FromBase64String(json.images[0]);
-
-                // Write it in the specified project output folder
-                using (FileStream imageFile = new FileStream(filename, FileMode.Create))
-                {
-#if UNITY_EDITOR
-                    AssetDatabase.StartAssetEditing();
-#endif
-                    yield return imageFile.WriteAsync(imageData, 0, imageData.Length);
-#if UNITY_EDITOR
-                    AssetDatabase.StopAssetEditing();
-                    AssetDatabase.SaveAssets();
-#endif
-                }
-
-                try
-                {
-                    // Read back the image into a texture
-                    if (File.Exists(filename))
-                    {
-                        Texture2D texture = new Texture2D(2, 2);
-                        texture.LoadImage(imageData);
-                        texture.Apply();
-
-                        LoadIntoImage(texture);
-                    }
-
-                    // Read the generation info back (only seed should have changed, as the generation picked a particular seed)
-                    if (json.info != "")
-                    {
-                        SDParamsOut info = JsonConvert.DeserializeObject<SDParamsOut>(json.info);
-
-                        // Read the seed that was used by Stable Diffusion to generate this result
-                        generatedSeed = info.seed;
-                    }
-                }
-                catch (Exception e)
-                {
-                    Debug.LogError(e.Message + "\n\n" + e.StackTrace);
-                }
-            }
-        }
-#if UNITY_EDITOR
-        EditorUtility.ClearProgressBar();
-#endif
-        generating = false;
-        yield return null;
-    }
-
-    /// <summary>
-    /// Load the texture into an Image or RawImage.
-    /// </summary>
-    /// <param name="texture">Texture to setup</param>
-    void LoadIntoImage(Texture2D texture)
-    {
-        try
-        {
-            // Find the image component
-            Image im = GetComponent<Image>();
-            if (im != null)
-            {
-                // Create a new Sprite from the loaded image
-                Sprite sprite = Sprite.Create(texture, new Rect(0, 0, texture.width, texture.height), Vector2.zero);
-
-                // Set the sprite as the source for the UI Image
-                im.sprite = sprite;
-            }
-            // If no image found, try to find a RawImage component
-            else
-            {
-                RawImage rim = GetComponent<RawImage>();
-                if (rim != null)
-                {
-                    rim.texture = texture;
-                }
-            }
-
-#if UNITY_EDITOR
-            if (!Application.isPlaying)
-            {
-                // Force Unity inspector to refresh with new asset
-                AssetDatabase.Refresh(ImportAssetOptions.ForceUpdate);
-                EditorApplication.QueuePlayerLoopUpdate();
-                EditorSceneManager.MarkAllScenesDirty();
-                EditorUtility.RequestScriptReload();
-            }
-#endif
-        }
-        catch (Exception e)
-        {
-            Debug.LogError(e.Message + "\n\n" + e.StackTrace);
-        }
-    }
-}
