+++ Assets/Stable-Diffusion-Unity-Integration/Scripts/StableDiffusionImage2Image.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f255497c431e7fc43ae59be84eb79cdd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
