--- Assets/Stable-Diffusion-Unity-Integration/Scripts/StableDiffusionMaterial.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 45733077a48ef55438bdf8ea5e658283
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
