+++ Assets/Stable-Diffusion-Unity-Integration/Scripts/StableDiffusionText2Image.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0beed6a09923ca540b45e4a3f055325f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
